+++ ui-plugins/muse-manager/src/index.js
@@ -13,3 +13,4 @@ plugin.register({
   route,
   reducer,
 });
+const a = (p) => p;