+++ workspace/packages/muse-dev-utils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ebay/muse-dev-utils
 
+## 1.0.23
+
+### Patch Changes
+
+- Dev testing.
+- Updated dependencies
+  - @ebay/muse-core@1.0.19
+  - @ebay/muse-express-middleware@1.0.21
+
 ## 1.0.22
 
 ### Patch Changes