+++ workspace/packages/muse-plugin-s3-storage/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ebay/muse-plugin-s3-storage
 
+## 1.0.18
+
+### Patch Changes
+
+- Dev testing.
+- Updated dependencies
+  - @ebay/muse-core@1.0.19
+
 ## 1.0.17
 
 ### Patch Changes